student profile

+++ database/migrations/2022_04_24_135517_add_academic_class_id_exam_schedules_table.php
@@ -14,7 +14,7 @@ class AddAcademicClassIdExamSchedulesTable extends Migration
     public function up()
     {
         Schema::table('exam_schedules', function (Blueprint $table) {
-            $table->integer('academic_class_id')->after('id');
+            $table->unsignedBigInteger('academic_class_id')->after('id');
         });
     }
